Let folks open a folder from outside Explorer
Now if you make a "ref" to explorer's component, you can call "openFolder(folder)". It will throw an exception if you pass a File which does not have the property "dir" set to true.
Added "root-drive" property. Now it's possible to set the root drive letter which was fixed and "C:\".